requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.33.0
psycopg2-binary==2.9.9
pandas==2.2.2
sqlparse==0.4.4
requests==2.31.0
faker==24.8.0
python-dotenv==1.0.1